feat: add 'Nothing' as a Powerwall export selection (#1098)
* Add support for 'Nothing' as a Powerwall export selection

Add support for specifying that the Powerwall system should not export any energy to the grid.

* Update test_select.py for Powerwall export option 'Nothing'

Add basic test for the Powerwall 'Nothing' export option.

* Corrected typo

* Temporarily comment out failing test

To see if it is this code that is causing the failure.

* Restoring original test_select.py

If test still breaks then it is because test was already broken.

* Fixed sseryion logic in test_select.py for new export rule case

One of the original options was incorrectly using assert_awaited_with_once and I had originally duplicated that rule. Changed the existing rule to use assert_awaited_with and added the new rule also using that.